+++ main/sched.c
@@ -316,7 +316,7 @@ void ast_sched_dump(const struct sched_context *con)
 			q->id,
 			q->callback,
 			q->data,
-			delta.tv_sec,
+			(long) delta.tv_sec,
 			(long int)delta.tv_usec);
 	}
 	ast_log(LOG_DEBUG, "=============================================================\n");
